format/mainPage_treeInfos 북마크 컴포넌트 세분화, 폴더구조 변경
