+++ src/components/common/BookMark/style.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components'
+
+export const Bookmark = styled.button`
+  width: 25px;
+  height: 24px;
+  background: #ffffff;
+  border: none;
+  padding: 0;
+  cursor: pointer;
+`
+
+export const IconTitle = styled.span`
+  font-size: #2c2c2c;
+  font-weight: 300;
+  font-size: 10px;
+  margin-top: 5px;
+`
